not finished

+++ students/views.py
@@ -1,9 +1,9 @@
-from django.shortcuts import render, redirect, get_object_or_404
+from django.shortcuts import render, redirect, reverse, get_object_or_404
 from .models import Student
+from .models import Group
 
 
-def home(request):
-    return render(request, 'index.html')
+
 
 def student_list(request):
     students = Student.objects.all()
@@ -13,12 +13,13 @@ def student_list(request):
 def student_create(request):
     if request.method == 'POST':
         full_name = request.POST.get('full_name')
-        group = request.POST.get('group')
+        group_id = request.POST.get('group')
         date_of_birth = request.POST.get('date_of_birth')
         student_phone = request.POST.get('student_phone')
         address = request.POST.get('address')
-        photo = request.POST.get('photo')
-        if full_name and group and date_of_birth and student_phone and address and photo:
+        photo = request.FILES.get('photo')
+        if full_name and group_id and date_of_birth and student_phone and address and photo:
+            group = get_object_or_404(Group, pk=group_id)
             Student.objects.create(
                 full_name=full_name,
                 group=group,
@@ -32,33 +33,36 @@ def student_create(request):
 
 def student_detail(request, pk):
     student = get_object_or_404(Student, pk=pk)
-    ctx = {'student', student}
+    ctx = {'student': student}
     return render(request, 'students/student-detail.html', ctx)
 
 def student_delete(request, pk):
     student = get_object_or_404(Student, pk=pk)
-    student.delete()
-    return redirect('students:list')
+    if request.method == 'POST':
+        student.delete()
+    ctx = {'student': student}
+    return redirect('students/student-delete-confirm.html', ctx)
 
 
 def student_update(request, pk):
     student = get_object_or_404(Student, pk=pk)
     if request.method == 'POST':
         full_name = request.POST.get('full_name')
-        group = request.POST.get('group')
+        group_id = request.POST.get('group')
         date_of_birth = request.POST.get('date_of_birth')
         student_phone = request.POST.get('student_phone')
         address = request.POST.get('address')
-        photo = request.FILE.get('photo')
-        if full_name and group and date_of_birth and student_phone and address and photo:
+        photo = request.FILES.get('photo')
+        if full_name and group_id and date_of_birth and student_phone and address and photo:
+            student.group = get_object_or_404(Group, pk=group_id)
             student.full_name = full_name
-            student.group = group
             student.date_of_birth = date_of_birth
             student.student_phone = student_phone
             student.address = address
-            student.photo = photo
+            if photo:
+                student.photo = photo
             student.save()
-            return redirect(student.get_detail_url())
+            return redirect(reverse('student:detail', args=[student.pk]))
     ctx = {'student': student}
     return render(request, 'students/student-form.html', ctx)
